Add folder structure in optimizers.
