+++ gqcp/include/Mathematical/Optimization/CMakeLists.txt
@@ -1,3 +1,4 @@
 add_subdirectory(Accelerator)
 add_subdirectory(Eigenproblem)
-add_subdirectory(NonLinear)
+add_subdirectory(Minimization)
+add_subdirectory(NonLinearEquation)
