+++ gqcp/include/Mathematical/Optimization/Minimization/BaseHessianModifier.hpp
