+++ gqcp/include/Mathematical/Optimization/Minimization/BaseMinimizer.hpp
