+++ gqcp/include/Mathematical/Optimization/NonLinearEquation/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(gqcp
+    PRIVATE
+        NewtonNLSystemOfEquationsSolver.hpp
+        step.hpp
+)
