+++ gqcp/src/Mathematical/Optimization/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory(Eigenproblem)
-add_subdirectory(NonLinear)
+add_subdirectory(Minimization)
+add_subdirectory(NonLinearEquation)
