+++ gqcp/src/Mathematical/Optimization/Minimization/BaseMinimizer.cpp
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-gqcp.  If not, see <http://www.gnu.org/licenses/>.
 // 
-#include "Mathematical/Optimization/NonLinear/BaseMinimizer.hpp"
+#include "Mathematical/Optimization/Minimization/BaseMinimizer.hpp"
 
 
 namespace GQCP {
