+++ gqcp/tests/Mathematical/Optimization/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_subdirectory(Eigenproblem)
-add_subdirectory(NonLinear)
+add_subdirectory(Minimization)
+add_subdirectory(NonLinearEquation)
 
 set(test_target_sources ${test_target_sources} PARENT_SCOPE)
