+++ gqcp/tests/Mathematical/Optimization/Minimization/CMakeLists.txt
@@ -1,7 +1,6 @@
 list(APPEND test_target_sources
     ${CMAKE_CURRENT_SOURCE_DIR}/IterativeIdentitiesHessianModifier_test.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/NewtonMinimizer_test.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/NewtonNLSystemOfEquationsSolver_test.cpp
 )
 
 set(test_target_sources ${test_target_sources} PARENT_SCOPE)
